backend/requirements.txt
Django>=4.2
djangorestframework
djoser
Pillow
psycopg2-binary
django-filter
